+++ ntex-util/src/services/variant.rs
@@ -143,23 +143,10 @@ macro_rules! variant_impl ({$mod_name:ident, $enum_type:ident, $srv_type:ident, 
             }).await
         }
 
-        async fn not_ready(&self) {
-            use std::{future::Future, pin::Pin};
-
-            let mut fut1 = ::std::pin::pin!(self.V1.not_ready());
-            $(let mut $T = ::std::pin::pin!(self.$T.not_ready());)+
-
-            ::std::future::poll_fn(|cx| {
-                if Pin::new(&mut fut1).poll(cx).is_ready() {
-                    return Poll::Ready(())
-                }
-
-                $(if Pin::new(&mut $T).poll(cx).is_ready() {
-                    return Poll::Ready(());
-                })+
-
-                Poll::Pending
-            }).await
+        fn poll(&self, cx: &mut std::task::Context<'_>) -> Result<(), Self::Error> {
+            self.V1.poll(cx)?;
+            $(self.$T.poll(cx)?;)+
+            Ok(())
         }
 
         async fn shutdown(&self) {
@@ -253,7 +240,6 @@ variant_impl_and!(VariantFactory7, VariantFactory8, V8, V8R, v8, (V2, V3, V4, V5
 #[cfg(test)]
 mod tests {
     use ntex_service::fn_factory;
-    use std::{future::poll_fn, future::Future, pin};
 
     use super::*;
 
@@ -307,16 +293,7 @@ mod tests {
         let service = factory.pipeline(&()).await.unwrap().clone();
         assert!(format!("{:?}", service).contains("Variant"));
 
-        let mut f = pin::pin!(service.not_ready());
-        let _ = poll_fn(|cx| {
-            if pin::Pin::new(&mut f).poll(cx).is_pending() {
-                Poll::Ready(())
-            } else {
-                Poll::Pending
-            }
-        })
-        .await;
-
+        assert!(crate::future::lazy(|cx| service.poll(cx)).await.is_ok());
         assert!(service.ready().await.is_ok());
         service.shutdown().await;